+++ CHANGELOG.md
@@ -11,6 +11,7 @@ Each release header has the following format:
 The title describes the overall update made with this release. The tag is a
 short marker that begins each of the related commit messages.
 
+ * Add __doc__ to describe test_whitespace()
 ## [0.0.3] / 2018-09-17 / Detail tweaks (details, DT)
 ### Changes
  * Change the date of Python 2.7's end of support, per pyclock.org